final project

+++ WebApplication1/Controllers/PlaylistsController.cs
@@ -19,13 +19,11 @@ namespace WebApplication1.Controllers
             _context = context;
         }
 
-        // GET: Playlists
         public async Task<IActionResult> Index()
         {
               return View(await _context.Playlist.ToListAsync());
         }
 
-        // GET: Playlists/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null || _context.Playlist == null) return NotFound();
@@ -41,15 +39,11 @@ namespace WebApplication1.Controllers
             return View(playlist);
         }
 
-        // GET: Playlists/Create
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Playlists/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Img,TracksCount")] Playlist playlist)
@@ -64,7 +58,6 @@ namespace WebApplication1.Controllers
             return View(playlist);
         }
 
-        // GET: Playlists/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Playlist == null) return NotFound();
@@ -79,73 +72,50 @@ namespace WebApplication1.Controllers
             return View(playlist);
         }
 
-        // POST: Playlists/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Img,TracksCount")] Playlist playlist)
         {
-            if (id != playlist.Id)
-            {
-                return NotFound();
-            }
+            if (id != playlist.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
                 try
                 {
+                    var tracksById = _context.Track.Where(t => t.PlaylistId == playlist.Id);
+                    playlist.TracksCount = tracksById.Count();
                     _context.Update(playlist);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!PlaylistExists(playlist.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
+                    if (!PlaylistExists(playlist.Id)) return NotFound();
                         throw;
-                    }
                 }
                 return RedirectToAction(nameof(Index));
             }
             return View(playlist);
         }
 
-        // GET: Playlists/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Playlist == null)
-            {
-                return NotFound();
-            }
+            if (id == null || _context.Playlist == null) return NotFound();
 
             var playlist = await _context.Playlist
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (playlist == null)
-            {
-                return NotFound();
-            }
+            if (playlist == null) return NotFound();
 
             return View(playlist);
         }
 
-        // POST: Playlists/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Playlist == null)
-            {
-                return Problem("Entity set 'WebApplication1Context.Playlist'  is null.");
-            }
+            if (_context.Playlist == null) return Problem("Entity set 'WebApplication1Context.Playlist'  is null.");
+
             var playlist = await _context.Playlist.FindAsync(id);
-            if (playlist != null)
-            {
-                _context.Playlist.Remove(playlist);
-            }
+            if (playlist != null) _context.Playlist.Remove(playlist);
 
             var tracksById = _context.Track.Where(t => t.PlaylistId == playlist.Id);
             foreach(var track in tracksById)
@@ -164,12 +134,12 @@ namespace WebApplication1.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> AddTrack(int id) {
-
-            var playlist = await _context.Playlist.FindAsync(id);
+        public async Task<IActionResult> AddTrack(int id) 
+        {
             var track = await _context.Track.FindAsync(Constants.trackId);
+            var playlist = await _context.Playlist.FindAsync(id);           
 
-            if (track.PlaylistId != playlist.Id)
+            if (track.PlaylistId == 0)
             {
                 track.PlaylistId = playlist.Id;
                 playlist.TracksCount += 1;
